results_analysis/requirements.txt
ailment==8.20.1.7
angr==8.20.1.7
archinfo==8.20.1.7
Babel==2.8.0
backcall==0.2.0
bitstring==3.1.6
cachetools==4.1.0
capstone==4.0.1
cffi==1.14.0
claripy==8.20.1.7
cle==8.20.1.7
Click==7.0
configparser==3.7.4
cycler==0.10.0
decorator==4.4.2
dnspython==1.16.0
dpkt==1.9.2
Flask==1.0.2
Flask-Babel==1.0.0
Flask-Cors==3.0.7
Flask-Table==0.5.0
future==0.18.2
gitdb==4.0.4
GitPython==3.1.1
graphviz==0.14.1
ipdb==0.13.3
ipython==7.16.1
ipython-genutils==0.2.0
itanium-demangler==1.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.10
kiwisolver==1.2.0
MarkupSafe==1.1.0
matplotlib==3.2.1
minidump==0.0.10
mulpyplexer==0.8
mysql-connector-python==8.0.19
networkx==2.4
numpy==1.18.4
pandas==1.1.5
parso==0.7.1
patsy==0.5.1
pefile==2019.4.18
pexpect==4.8.0
pickleshare==0.7.5
plumbum==1.6.9
prettytable==0.7.2
progressbar2==3.51.0
prompt-toolkit==3.0.7
protobuf==3.6.1
psutil==5.7.0
ptyprocess==0.6.0
pycparser==2.20
pydot==1.4.1
pyelftools==0.26
Pygments==2.7.1
pygraphviz==1.6
pyparsing==2.4.7
PySMT==0.9.0
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2020.1
pyvex==8.20.1.7
r2pipe==1.4.2
rpyc==4.1.5
scipy==1.5.2
seaborn==0.11.1
six==1.14.0
smmap==3.0.2
sortedcontainers==2.1.0
statsmodels==0.12.2
traitlets==4.3.3
unicorn==1.0.1
wcwidth==0.2.5
Werkzeug==0.14.1
z3-solver==4.8.7.0
